refactor: :recycle: Separate eval sets logic and competition logic
